+++ saleor/graphql/checkout/tests/mutations/test_checkout_complete.py
@@ -435,28 +435,28 @@ def test_checkout_complete_fails_with_invalid_tax_app(
 
     checkout.refresh_from_db()
     assert checkout.price_expiration == timezone.now() + settings.CHECKOUT_PRICES_TTL
-    assert checkout.tax_error == "Empty tax data."
+    assert checkout.tax_error == "Configured tax app doesn't exist."
 
 
 @freeze_time()
 @override_settings(PLUGINS=["saleor.plugins.webhook.plugin.WebhookPlugin"])
-@mock.patch("saleor.checkout.calculations.validate_tax_data")
 @mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
 def test_checkout_complete_calls_correct_tax_app(
     mock_request,
-    mock_validate_tax_data,
     user_api_client,
     checkout_without_shipping_required,
     channel_USD,
     address,
     tax_app,
-    tax_data_response,  # noqa: F811
+    tax_data_response_factory,  # noqa: F811
     settings,
 ):
     # given
-    mock_request.return_value = tax_data_response
-    mock_validate_tax_data.return_value = False
     checkout = checkout_without_shipping_required
+    mock_request.return_value = tax_data_response_factory(
+        lines_length=checkout.lines.count()
+    )
+
     checkout.billing_address = address
     checkout.price_expiration = timezone.now()
     checkout.metadata_storage.store_value_in_metadata(items={"accepted": "true"})
@@ -553,24 +553,25 @@ def test_checkout_complete_calls_failing_plugin(
 
 @freeze_time()
 @override_settings(PLUGINS=["saleor.plugins.webhook.plugin.WebhookPlugin"])
-@mock.patch("saleor.checkout.calculations.validate_tax_data")
 @mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
 def test_checkout_complete_calls_correct_force_tax_calculation_when_tax_error_was_saved(
     mock_request,
-    mock_validate_tax_data,
     user_api_client,
     checkout_without_shipping_required,
     channel_USD,
     address,
     tax_app,
-    tax_data_response,  # noqa: F811
+    tax_data_response_factory,  # noqa: F811
     settings,
 ):
     # given
-    mock_request.return_value = tax_data_response
-    mock_validate_tax_data.return_value = False
 
     checkout = checkout_without_shipping_required
+
+    mock_request.return_value = tax_data_response_factory(
+        lines_length=checkout.lines.count()
+    )
+
     checkout.billing_address = address
     checkout.price_expiration = (
         timezone.now() + settings.CHECKOUT_PRICES_TTL + timezone.timedelta(hours=1)