+++ saleor/plugins/webhook/plugin.py
@@ -8,6 +8,8 @@ from typing import TYPE_CHECKING, Any, Final, Optional, Union
 
 import graphene
 from django.conf import settings
+from django.template.defaultfilters import truncatechars
+from pydantic import ValidationError
 
 from ...app.models import App
 from ...channel.models import Channel
@@ -16,7 +18,7 @@ from ...checkout.models import Checkout
 from ...core import EventDeliveryStatus
 from ...core.models import EventDelivery
 from ...core.notify import NotifyEventType
-from ...core.taxes import TaxData, TaxType
+from ...core.taxes import TAX_ERROR_FIELD_LENGTH, TaxData, TaxDataError, TaxType
 from ...core.utils import build_absolute_uri
 from ...core.utils.json_serializer import CustomJsonEncoder
 from ...csv.notifications import get_default_export_payload
@@ -29,6 +31,7 @@ from ...graphql.webhook.utils import (
     get_pregenerated_subscription_payload,
     get_subscription_query_hash,
 )
+from ...order.models import Order
 from ...payment import PaymentError, TransactionKind
 from ...payment.interface import (
     GatewayResponse,
@@ -107,7 +110,7 @@ from ...webhook.transport.shipping import (
     parse_list_shipping_methods_response,
 )
 from ...webhook.transport.synchronous.transport import (
-    trigger_all_webhooks_sync,
+    trigger_taxes_all_webhooks_sync,
     trigger_transaction_request,
     trigger_webhook_sync,
     trigger_webhook_sync_if_not_cached,
@@ -158,7 +161,6 @@ if TYPE_CHECKING:
 # time labels were valid for when checking documentation for the carriers
 # (FedEx, UPS, TNT, DHL).
 CACHE_TIME_SHIPPING_LIST_METHODS_FOR_CHECKOUT: Final[int] = 3600 * 12
-
 
 logger = logging.getLogger(__name__)
 
@@ -3387,9 +3389,10 @@ class WebhookPlugin(BasePlugin):
         event_type: str,
         app_identifier: str,
         payload_gen: Callable,
+        expected_lines_count: int,
         subscriptable_object=None,
         pregenerated_subscription_payloads: dict | None = None,
-    ):
+    ) -> TaxData:
         if pregenerated_subscription_payloads is None:
             pregenerated_subscription_payloads = {}
         app = (
@@ -3402,14 +3405,14 @@ class WebhookPlugin(BasePlugin):
             .first()
         )
         if app is None:
-            logger.warning("Configured tax app doesn't exists.")
-            return None
+            msg = "Configured tax app doesn't exist."
+            logger.warning(msg)
+            raise TaxDataError(msg)
         webhook = get_webhooks_for_event(event_type, apps_ids=[app.id]).first()
         if webhook is None:
-            logger.warning(
-                "Configured tax app's webhook for checkout taxes doesn't exists."
-            )
-            return None
+            msg = "Configured tax app's webhook for taxes calculation doesn't exists."
+            logger.warning(msg)
+            raise TaxDataError(msg)
 
         request_context = initialize_request(
             self.requestor,
@@ -3431,7 +3434,19 @@ class WebhookPlugin(BasePlugin):
             requestor=self.requestor,
             pregenerated_subscription_payload=pregenerated_subscription_payload,
         )
-        return parse_tax_data(response)
+        try:
+            tax_data = parse_tax_data(response, expected_lines_count)
+        except ValidationError as e:
+            errors = e.errors()
+            logger.warning(
+                "Webhook response for event %s is invalid: %s",
+                event_type,
+                str(e),
+                extra={"errors": errors},
+            )
+            error_msg = truncatechars(str(e), TAX_ERROR_FIELD_LENGTH)
+            raise TaxDataError(error_msg, errors=errors) from e
+        return tax_data
 
     def get_taxes_for_checkout(
         self,
@@ -3440,10 +3455,11 @@ class WebhookPlugin(BasePlugin):
         app_identifier,
         previous_value,
         pregenerated_subscription_payloads: dict | None = None,
-    ) -> Optional["TaxData"]:
+    ) -> TaxData | None:
         if pregenerated_subscription_payloads is None:
             pregenerated_subscription_payloads = {}
         event_type = WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES
+        lines_count = len(lines)
         if app_identifier:
             return self.__run_tax_webhook(
                 event_type,
@@ -3451,16 +3467,20 @@ class WebhookPlugin(BasePlugin):
                 lambda: generate_checkout_payload_for_tax_calculation(
                     checkout_info, lines
                 ),
+                lines_count,
                 checkout_info.checkout,
                 pregenerated_subscription_payloads=pregenerated_subscription_payloads,
             )
-        return trigger_all_webhooks_sync(
+        # This is deprecated flow, kept to maintain backward compatibility.
+        # In Saleor 4.0 `tax_app_identifier` should be required and the flow should
+        # be dropped.
+        return trigger_taxes_all_webhooks_sync(
             event_type,
             lambda: generate_checkout_payload_for_tax_calculation(
                 checkout_info,
                 lines,
             ),
-            parse_tax_data,
+            lines_count,
             checkout_info.checkout,
             self.requestor,
             pregenerated_subscription_payloads=pregenerated_subscription_payloads,
@@ -3468,19 +3488,24 @@ class WebhookPlugin(BasePlugin):
 
     def get_taxes_for_order(
         self, order: "Order", app_identifier, previous_value
-    ) -> Optional["TaxData"]:
+    ) -> TaxData | None:
         event_type = WebhookEventSyncType.ORDER_CALCULATE_TAXES
+        lines_count = order.lines.count()
         if app_identifier:
             return self.__run_tax_webhook(
                 event_type,
                 app_identifier,
                 lambda: generate_order_payload_for_tax_calculation(order),
+                lines_count,
                 order,
             )
-        return trigger_all_webhooks_sync(
+        # This is deprecated flow, kept to maintain backward compatibility.
+        # In Saleor 4.0 `tax_app_identifier` should be required and the flow should
+        # be dropped.
+        return trigger_taxes_all_webhooks_sync(
             WebhookEventSyncType.ORDER_CALCULATE_TAXES,
             lambda: generate_order_payload_for_tax_calculation(order),
-            parse_tax_data,
+            lines_count,
             order,
             self.requestor,
         )